backend/requirements.txt
# Backend dependencies for Skill-Bridge Career Navigator

# FastAPI and web framework dependencies
fastapi==0.115.0
uvicorn[standard]==0.24.0
pydantic==2.9.0

# HTTP and API dependencies
httpx==0.25.2
requests==2.31.0
urllib3>=1.26.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1

# Development and utility dependencies
python-multipart==0.0.9

# AI/ML dependencies
google-generativeai==0.8.0

# Additional utilities
python-dotenv==1.0.0